pyproject.toml
[project]
name = "evalaite"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "google-genai",
    "sqlalchemy",
    "pdfplumber",
    "pillow",
    "python-dotenv",
    "jinja2",
    "aiofiles",
    "pydantic-settings",
]
